allow app for many users

+++ src/routes/account/+page.server.ts
@@ -1,24 +1,17 @@
-import { fail, redirect, type Actions } from '@sveltejs/kit'
+import { error, fail, redirect, type Actions } from '@sveltejs/kit'
 import bcrypt from 'bcrypt'
 import { db, query } from '$lib/server/db'
-import type { PageServerLoad } from './$types'
-import {
-	COOKIE_OPTIONS,
-	COOKIE_USERNAME,
-	MINIMAL_PASSWORD_LENGTH,
-} from '$lib/server/config'
+import { MINIMAL_PASSWORD_LENGTH } from '$lib/server/config'
 import { SUPPORTED_LANGUAGES, ts, type Lang } from '$lib/translations/main'
 import { verify_entries } from '$lib/utils'
 import { encrypt_entry } from '$lib/server/encryption'
 import { get_language, set_language_cookie } from '$lib/translations/request'
-
-export const load: PageServerLoad = async (event) => {
-	const username = event.cookies.get(COOKIE_USERNAME) ?? ''
-	return { username }
-}
+import { delete_auth_cookie, set_auth_cookie } from '$lib/server/auth'
 
 export const actions: Actions = {
 	lang: async (event) => {
+		const user = event.locals.user
+		if (!user) error(401, 'Unauthorized')
 		const form = await event.request.formData()
 		const lang_option = form.get('lang') as string
 		if ((SUPPORTED_LANGUAGES as readonly string[]).includes(lang_option)) {
@@ -27,13 +20,17 @@ export const actions: Actions = {
 	},
 
 	password: async (event) => {
+		const user = event.locals.user
+		if (!user) error(401, 'Unauthorized')
+
 		const lang = get_language(event.cookies)
 		const form = await event.request.formData()
 		const current_password = form.get('current_password') as string
 		const new_password = form.get('new_password') as string
 
 		const { rows, success } = await query<{ password_hash: string }>(
-			'SELECT password_hash FROM users WHERE id = 1',
+			'SELECT password_hash FROM users WHERE id = ?',
+			[user.id],
 		)
 
 		if (!success || !rows.length) {
@@ -43,9 +40,9 @@ export const actions: Actions = {
 			})
 		}
 
-		const user = rows[0]
+		const { password_hash } = rows[0]
 
-		const current_is_correct = await bcrypt.compare(current_password, user.password_hash)
+		const current_is_correct = await bcrypt.compare(current_password, password_hash)
 		if (!current_is_correct) {
 			return fail(401, {
 				type: 'password',
@@ -60,11 +57,11 @@ export const actions: Actions = {
 			})
 		}
 
-		const password_hash = await bcrypt.hash(new_password, 10)
+		const new_password_hash = await bcrypt.hash(new_password, 10)
 
 		const { success: update_success } = await query(
-			'UPDATE users SET password_hash = ? WHERE id = 1',
-			[password_hash],
+			'UPDATE users SET password_hash = ? WHERE id = ?',
+			[new_password_hash, user.id],
 		)
 
 		if (!update_success) {
@@ -81,6 +78,9 @@ export const actions: Actions = {
 	},
 
 	username: async (event) => {
+		const user = event.locals.user
+		if (!user) error(401, 'Unauthorized')
+
 		const lang = get_language(event.cookies)
 		const form = await event.request.formData()
 		const username = form.get('username') as string
@@ -92,8 +92,9 @@ export const actions: Actions = {
 			})
 		}
 
-		const { success } = await query('UPDATE users SET username = ? WHERE id = 1', [
+		const { success } = await query('UPDATE users SET username = ? WHERE id = ?', [
 			username,
+			user.id,
 		])
 
 		if (!success) {
@@ -103,7 +104,9 @@ export const actions: Actions = {
 			})
 		}
 
-		event.cookies.set(COOKIE_USERNAME, username, COOKIE_OPTIONS)
+		user.username = username
+
+		set_auth_cookie(event, user)
 
 		return {
 			type: 'username',
@@ -112,6 +115,9 @@ export const actions: Actions = {
 	},
 
 	backup: async (event) => {
+		const user = event.locals.user
+		if (!user) error(401, 'Unauthorized')
+
 		const lang = get_language(event.cookies)
 		const form = await event.request.formData()
 		const file = form.get('file')
@@ -157,13 +163,14 @@ export const actions: Actions = {
 
 			for (const entry of encoded_entries) {
 				await tx.execute({
-					sql: 'INSERT INTO entries (id, date, title_enc, content_enc, thanks_enc) VALUES (?,?,?,?,?)',
+					sql: 'INSERT INTO entries (id, date, title_enc, content_enc, thanks_enc, user_id) VALUES (?,?,?,?,?,?)',
 					args: [
 						entry.id,
 						entry.date,
 						entry.title_enc,
 						entry.content_enc,
 						entry.thanks_enc,
+						user.id,
 					],
 				})
 			}
@@ -181,6 +188,9 @@ export const actions: Actions = {
 	},
 
 	delete: async (event) => {
+		const user = event.locals.user
+		if (!user) error(401, 'Unauthorized')
+
 		const lang = get_language(event.cookies)
 		const form = await event.request.formData()
 		const user_yes = form.get('yes') as string
@@ -193,22 +203,16 @@ export const actions: Actions = {
 			})
 		}
 
-		const tx = await db.transaction('write')
+		const { success } = await query('DELETE FROM users WHERE id = ?', [user.id])
 
-		try {
-			await tx.execute('DELETE FROM entries')
-			await tx.execute('DELETE FROM users')
-			await tx.execute('DELETE FROM devices')
-			await tx.commit()
-		} catch (_) {
+		if (!success) {
 			return fail(500, {
 				type: 'delete',
 				error: ts('error.database', lang),
 			})
 		}
 
-		event.cookies.delete('jwt', { path: '/' })
-		event.cookies.delete('username', { path: '/' })
+		delete_auth_cookie(event)
 
 		return redirect(302, '/')
 	},
